Change test selection logic and cleanup
The previous test selection logic would scan all tests and filter some
out, instead switch to per module level test selection logic.

Cleanup:
* Apply black and isort etc.
* Drop logic for coverage for now to simplify the code base
* Simplify UTC logic
* Set no_summary and remove summary_ functions

In the future it would be nice to update the selection logic to work per test
as well. Didn't figure out a quick way to solve that one, so will have
to revisit in a future pass.